refactor: modularize main entry
