+++ chat_utils.ts
@@ -1,0 +1,23 @@
+import {GoogleGenAI, mcpToTool} from '@google/genai';
+import {Client} from '@modelcontextprotocol/sdk/client/index.js';
+import {SYSTEM_INSTRUCTIONS} from './system_instructions.js';
+
+export function createAiChat(mcpClient: Client, geminiApiKey: string) {
+  const ai = new GoogleGenAI({
+    apiKey: geminiApiKey,
+  });
+  return ai.chats.create({
+    model: 'gemini-2.5-flash',
+    config: {
+      systemInstruction: SYSTEM_INSTRUCTIONS,
+      tools: [mcpToTool(mcpClient)],
+    },
+  });
+}
+
+export function camelCaseToDash(str: string): string {
+  return str
+    .replace(/([a-z])([A-Z])/g, '$1-$2')
+    .replace(/([A-Z])([A-Z][a-z])/g, '$1-$2')
+    .toLowerCase();
+}
