+++ send_message_handler.ts
@@ -1,0 +1,171 @@
+import {ChatState, MapApp, marked} from './map_app.js';
+import {camelCaseToDash} from './chat_utils.js';
+
+export function createSendMessageHandler(mapApp: MapApp, aiChat: any) {
+  return async (input: string, role: string) => {
+    console.log('sendMessageHandler', input, role);
+
+    const {thinkingElement, textElement, thinkingContainer} = mapApp.addMessage(
+      'assistant',
+      '',
+    );
+
+    mapApp.setChatState(ChatState.GENERATING);
+    textElement.innerHTML = '...'; // Initial placeholder
+
+    let newCode = '';
+    let thoughtAccumulator = '';
+
+    try {
+      // Outer try for overall message handling including post-processing
+      try {
+        // Inner try for AI interaction and message parsing
+        const stream = await aiChat.sendMessageStream({message: input});
+
+        for await (const chunk of stream) {
+          for (const candidate of chunk.candidates ?? []) {
+            for (const part of candidate.content?.parts ?? []) {
+              if (part.functionCall) {
+                console.log(
+                  'FUNCTION CALL:',
+                  part.functionCall.name,
+                  part.functionCall.args,
+                );
+                const mcpCall = {
+                  name: camelCaseToDash(part.functionCall.name!),
+                  arguments: part.functionCall.args,
+                };
+
+                const explanation =
+                  'Calling function:\n```json\n' +
+                  JSON.stringify(mcpCall, null, 2) +
+                  '\n```';
+                const {textElement: functionCallText} = mapApp.addMessage(
+                  'assistant',
+                  '',
+                );
+                functionCallText.innerHTML = await marked.parse(explanation);
+              }
+
+              if (part.thought) {
+                mapApp.setChatState(ChatState.THINKING);
+                thoughtAccumulator += ' ' + part.thought;
+                thinkingElement.innerHTML =
+                  await marked.parse(thoughtAccumulator);
+                if (thinkingContainer) {
+                  thinkingContainer.classList.remove('hidden');
+                  thinkingContainer.setAttribute('open', 'true');
+                }
+              } else if (part.text) {
+                mapApp.setChatState(ChatState.EXECUTING);
+                newCode += part.text;
+                textElement.innerHTML = await marked.parse(newCode);
+              }
+              mapApp.scrollToTheEnd();
+            }
+          }
+        }
+      } catch (e: unknown) {
+        // Catch for AI interaction errors.
+        console.error('GenAI SDK Error:', e);
+        let baseErrorText: string;
+
+        if (e instanceof Error) {
+          baseErrorText = e.message;
+        } else if (typeof e === 'string') {
+          baseErrorText = e;
+        } else if (
+          e &&
+          typeof e === 'object' &&
+          'message' in e &&
+          typeof (e as {message: unknown}).message === 'string'
+        ) {
+          baseErrorText = (e as {message: string}).message;
+        } else {
+          try {
+            // Attempt to stringify complex objects, otherwise, simple String conversion.
+            baseErrorText = `Unexpected error: ${JSON.stringify(e)}`;
+          } catch (stringifyError) {
+            baseErrorText = `Unexpected error: ${String(e)}`;
+          }
+        }
+
+        let finalErrorMessage = baseErrorText; // Start with the extracted/formatted base error message.
+
+        // Attempt to parse a JSON object from the baseErrorText, as some SDK errors embed details this way.
+        // This is useful if baseErrorText itself is a string containing JSON.
+        const jsonStartIndex = baseErrorText.indexOf('{');
+        const jsonEndIndex = baseErrorText.lastIndexOf('}');
+
+        if (jsonStartIndex > -1 && jsonEndIndex > jsonStartIndex) {
+          const potentialJson = baseErrorText.substring(
+            jsonStartIndex,
+            jsonEndIndex + 1,
+          );
+          try {
+            const sdkError = JSON.parse(potentialJson);
+            let refinedMessageFromSdkJson: string | undefined;
+
+            // Check for common nested error structures (e.g., sdkError.error.message)
+            // or a direct message (sdkError.message) in the parsed JSON.
+            if (
+              sdkError &&
+              typeof sdkError === 'object' &&
+              sdkError.error && // Check if 'error' property exists and is truthy
+              typeof sdkError.error === 'object' && // Check if 'error' property is an object
+              typeof sdkError.error.message === 'string' // Check for 'message' string within 'error' object
+            ) {
+              refinedMessageFromSdkJson = sdkError.error.message;
+            } else if (
+              sdkError &&
+              typeof sdkError === 'object' && // Check if sdkError itself is an object
+              typeof sdkError.message === 'string' // Check for a direct 'message' string on sdkError
+            ) {
+              refinedMessageFromSdkJson = sdkError.message;
+            }
+
+            if (refinedMessageFromSdkJson) {
+              finalErrorMessage = refinedMessageFromSdkJson; // Update if JSON parsing yielded a more specific message
+            }
+          } catch (parseError) {
+            // If parsing fails, finalErrorMessage remains baseErrorText.
+            console.warn(
+              'Could not parse potential JSON from error message; using base error text.',
+              parseError,
+            );
+          }
+        }
+
+        const {textElement: errorTextElement} = mapApp.addMessage('error', '');
+        errorTextElement.innerHTML = await marked.parse(
+          `Error: ${finalErrorMessage}`,
+        );
+      }
+
+      // Post-processing logic (now inside the outer try)
+      if (thinkingContainer && thinkingContainer.hasAttribute('open')) {
+        if (!thoughtAccumulator) {
+          thinkingContainer.classList.add('hidden');
+        }
+        thinkingContainer.removeAttribute('open');
+      }
+
+      if (
+        textElement.innerHTML.trim() === '...' ||
+        textElement.innerHTML.trim().length === 0
+      ) {
+        const hasFunctionCallMessage = mapApp.messages.some((el) =>
+          el.innerHTML.includes('Calling function:'),
+        );
+        if (!hasFunctionCallMessage) {
+          textElement.innerHTML = await marked.parse('Done.');
+        } else if (textElement.innerHTML.trim() === '...') {
+          textElement.innerHTML = '';
+        }
+      }
+    } finally {
+      // Finally for the outer try, ensures chat state is reset
+      mapApp.setChatState(ChatState.IDLE);
+    }
+  };
+}
